Major Changes Made
For the Ontology Embedding the OWL2Vec-Star-master folder is added to our main folder which is R&S_Semantics folder and all the work which was done OWL2Vec-Star-master folder is now shown in main folder. And the first task which is creating 2 configuration has been made in binary and in text format. And both of these files are in OWL2Vec-Star-master folder.